Split form tests by widget ownership
