+++ packages/flutter/test/material/dropdown_form_field_test.dart
@@ -1529,4 +1529,73 @@ void main() {
       SystemMouseCursors.cell,
     );
   });
+
+  testWidgets('allows collecting and updating values from all field types', (
+    WidgetTester tester,
+  ) async {
+    final formKey = GlobalKey<FormState>();
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Form(
+            key: formKey,
+            child: Column(
+              children: <Widget>[
+                TextFormField(key: const ValueKey('name'), initialValue: 'Name'),
+                TextFormField(key: const ValueKey('email'), initialValue: 'Email'),
+                FormField<int>(
+                  key: const ValueKey('age'),
+                  initialValue: 18,
+                  builder: (field) => const SizedBox.shrink(),
+                ),
+                DropdownButtonFormField<String>(
+                  key: const ValueKey('animal'),
+                  initialValue: 'cat',
+                  items: const [
+                    DropdownMenuItem(value: 'dog', child: SizedBox.shrink()),
+                    DropdownMenuItem(value: 'cat', child: SizedBox.shrink()),
+                    DropdownMenuItem(value: 'bird', child: SizedBox.shrink()),
+                  ],
+                  onChanged: (_) {},
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+
+    Map<String, dynamic> collectData() {
+      return {
+        for (final field in formKey.currentState!.fields)
+          if (field.widget.key case ValueKey<String>(:final value)) value: field.value,
+      };
+    }
+
+    expect(collectData(), <String, Object>{
+      'name': 'Name',
+      'email': 'Email',
+      'age': 18,
+      'animal': 'cat',
+    });
+
+    FormFieldState<T> field<T>(String key) => formKey.currentState!.fields
+        .whereType<FormFieldState<T>>()
+        .singleWhere((f) => f.widget.key == ValueKey(key));
+
+    field<String>('name').didChange('New Name');
+    field<String>('email').didChange('new@email.com');
+    field<int>('age').didChange(30);
+    field<String>('animal').didChange('dog');
+
+    await tester.pump();
+
+    expect(collectData(), <String, Object>{
+      'name': 'New Name',
+      'email': 'new@email.com',
+      'age': 30,
+      'animal': 'dog',
+    });
+  });
 }
